hmsbackend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
Faker==26.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
